Switch tests away from enzyme.mount (#7796)
This switches all tests in `components/higher-order/with-context/test/index.js` from using enzyme.mount to `React.TestRenderer`.  This is because `enzyme` does not fully support React 16.3+ (and movement to do so is really slow). This will fix issues with breakage due to the enzyme incompatibility as components receive React 16.3+ features (such as `forwardRef` usage in #7557).