Switch away from using enzyme.mount in tests (components/higher-order/with-filters/test/index.js) (#7793)
* switch away from using enzyme.mount in tests

This switches all tests in `components/higher-order/with-filters/test/index.js` from using enzyme.mount to `React.TestUtilities`.  This is because `enzyme` does not fully support React 16.3+ (and movement to do so is really slow). This will fix issues with breakage due to the enzyme incompatibility as components receive React 16.3+ features (such as `forwardRef` usage in #7557).

* switch test to be against root node.

* fix linting issues

* remove unnecessary unmount