feature: Add ability for themes to configure font sizes; Extract fontSize logic from paragraph (#6628)
Extracts the logic to manage font sizes in a similar way to what we used to abstract the color logic.
Refactors the styles, so they are not specific to paragraphs. Now, other blocks can make use of the same classes.
The class names were renamed from the format is-large-text to has-large-font-size, this makes the names consistent with the names used in color classes. The new name makes clear that the primary style set is the font-size.
Font sizes are now read from editor settings instead of being a constant in the paragraph block.