Block Movers: Hide movers if there are no blocks before and after. (#8011)
If there are no blocks before or after the group of selected blocks, it is not possible to move the blocks anywhere.
In that case, we should not show the movers. Not showing the movers for this cases makes the UI cleaner in nested contexts as discovered in https://github.com/WordPress/gutenberg/pull/7414.