Fix issues with drag and drop over text (#9071)
* Fix issues with drag and drop over text

This PR is a followup to #8752, and fixes #8115.

It is the situation in `master` right now that every single block has instructional overlay text when dragging files from the desktop over it. This also means very short paragraphs, separators, lots of different blocks that can neither fit the upload message, nor can really _accept_ any uploads. That last part is key, because we also have a different upload indicator — the blue horizontal line. This indicator shows _where the block you're about to drop will be inserted_. Because you can drop an image file on top of a paragraph, but it will show up either _before_ the paragraph, or _after_ the paragraph, not _inside_ it.

Therefore, the upload message should not be present on top of every single block. In fact it should be present only on placeholders and galleries, where dropping a file on the block itself really _does_ insert the dropped file there.

* Avoid showing the drop zone text on paragraph Block DropZoneswq