Update icons for Paragraph, Heading and Subheading for clarity (#9152)
* Update icons for Paragraph, Heading and Subheading for clarity

This PR changes the Paragraph heading to an outline version of the paragraph symbol. This is done to fix the confusion where the text-icon looks confusing next to alignments: https://mobile.twitter.com/webmatros/status/1030816482614235137

By using the Paragraph icon, it shows what the block is.

This PR also adds new icons for Heading and Subheading, a capital H and smallcap italic H respectively. This is inspired by the same reasons as the "P for Paragraph" argument, H for Heading. Subheading was changed additionally because its previous visual language could also be confused with alignments.

* Remove fill that wasn't supposed to be there.