Update: Colors: Enhancements to the color mechanism ( renaming, logic updates).
The color mechanism suffered many changes since it was created.  Namely the fact that now the identified for colors is not the name but a slug. 
This PR applies some changes to improve the code quality.
It also removes the deprecation logic that should be removed in this version.

It applies some changes to make sure withColors now passed all the information contained in the color object as set by the themes to the blocks using withColors HOC.

The breaking change is that now the color code passed in objects by withColors is named as "color" instead of "value".
That change was made because of two reasons:
- To be consistent with the withFontSizes HOC which name the font size value as size.
- To be consistent with the color object that themes set, which names the color code as color: