Rich Text: Indicate which text will be turned into a link (#8807)
* Rich Text: Indicate which text will be turned into a link

When inserting a link, the text selection disappears when the focus
changes into the URLInput text field. This makes it hard to tell which
text will be turned into a link.

The Classic Editor solves this by inserting a placeholder <a> element,
which is the approach that we borrow here.

* Add E2E tests for creating, editing and removing links

* Use an `if` instead of a `switch` in getFormatValue()

A `switch` is overkill when there's only one format value that acts
differently.

* Rename Link E2E tests

Renames managing-links → links and makes the test descriptions read like
English sentences.