Automatically create a link when selected text is a URL (#9067)
* Automatically create a link when selected text is a URL

When a URL is selected and the Link button is clicked or the Link
keyboard shortcut is pressed, a link with that URL set as its href
should be instantly created.

* Simplify isURL tests

* Fix URL link editing

Fixes a bug that prevented the user from selecting a link which has text
that looks like a URL and clicking Edit. We accomplish this by
distinguishing between editing a link (isEditingLink) and adding a new
link (isAddingLink).