Add dark editor style support (#9683)
* Add dark editor style support

This fixes #6451.

This PR enables a new command to themers. They can now add `add_theme_support( 'dark-theme' );` in order to ask the editing canvas to be friendly to dark themes.

What this does is add a body class to the editor, `is-dark-theme`, which enables already present color changes to the side UI and borders, that are friendly to dark themes.

* Update documentation.

* Update theme support docs.

* Use dark-editor-style instead.

Hopefully this will help prevent confusion.