Send post_id to the REST API in the ServerSideRender component within the Editor (#9323)
This commit adds a wrapper around the `ServerSideRender` component within the `@wordpress/editor` package that adds a `post_id` parameter to the REST API call. This ensures the global `$post` object is set properly. Previously, it was only set when rendering to the frontend, but not when rendering for the API call.