Try a simpler spacing setup for gallery items (#10027)
* Try a simpler spacing setup for gallery items

The stock gallery features white gutters between items. Currently in `master`, every item has an even 8px margin all around. We then compensate for this, left and right, of the gallery, to make sure items in the gallery align correctly with text width above and below the gallery.

This PR attempts a different approach. It sets only the right and bottom margin on gallery items, and then unsets the right margin on the rightmost item in each row.

This is a little more complex in the rules that are output, but it solves the issue presented here: https://github.com/WordPress/gutenberg/pull/9670#issuecomment-419388164

Let me know your thoughts.

* Clean up the mess.

Remove debug comments. Sorry, long day.

Clarify comments.

* Address feedback.

* Address feedback.

* Bugfix.

* Adjust gallery css layout math, and ensure margin is reset correctly at mobile breakpoints

Co-authored-by: Sofia Sousa <csofiamsousa@gmail.com>

* Fix odd rule

Props @sofiasousa

* Add fixed rule