Clean up variables, comments, reduce complexity (#10430)
* Clean up variables, comments, reduce complexity

This PR does a few things:

1. It replaces the $item-spacing variable with $grid-size, which is one small step towards a base8 based grid.
2. It cleans up comments in the variables file, fixes a few typos, and clarifies comments.
3. It retires a rarely used variable and makes the SCSS math that used it slightly more clear.

Aside from a good code review and testing, this needs a visual regression sanity check. The $grid-size variable is 8px, whereas $item-spacing was 10px. In my own testing, this caused no issues, but it would still be good to check a bit more in-depth. If some margins are now too small, we can consider a 12px variant, or jump directly to 16px.

* Fix typo.

* Address feedback.

* Fix SCSS.