Fix: Mobile: Custom color picker does not appear on mobile. (#10481)
On small sizes, the color picker was not appearing. This happened because its z-index was lower than the z-index of the sidebar on mobile.
This commit fixes this z-index problem.

I added 'not(.is-mobile)' to an existing z-index because the rule where it was used was inside a not(.is-mobile) and that was not being referenced making debugging of z-index problems harder.