Make possible to overwrite the globally defined colors in PanelColorSettings (#10457)
This commit makes it possible to overwrite the globally defined colors (colors arrays and disableCustomColors) in PanelColorSettings.

The overwrite can be global for all the color panels of the component or local per each settings panel.

This commit makes sure that we offer all the functionality that the combination of wp.editor.PanelColor and wp.components.PanelColor had. Making it possible to deprecate these components and unify the color UI components.