wp.data: Split store implementation out from registry. (#10289)
* data: Start split of registry with reducer

This splits out the implementation of `registerReducer` into a separate
file, with the intention of continuing the rest of the store
implementation. This will make room for other possible implementations
with a common interface.

* data: Move registerSelectors and registerActions

This moves the implementation of registerSelectors and registerActions
to the namespace-store implementation file.

* data: Untangle registerResolvers

This untangles registerResolvers, puts more code into the
namespace-store, and organizes the functions upon which it depends.

* data: Centralize namespace registration

This commit centralizes the logic for namespace registration, in order
to support `registerStore` better.
`register[Reducer|Actions|Selectors|Resolvers]` now are viewed as
incomplete calls to `registerStore`

* data: Add registerGenericStore

This adds the registerGenericStore function and removes the `namespaces`
object in favor of a more generic `stores` object.

* data: Use store.subscribe() function

* data: Make registerGenericStore public + add tests

This makes `registry.registerGenericStore` accessible from the result of
`createRegistry`, and adds tests for the functionality of it.

* data: Add registerGenericStore to readme

This adds information about `registerGenericStore` to the readme with
two examples.

* data: Export registerGenericStore from index

This adds an export for `registerGenericStore` for the defaultRegistry.
While the defaultRegistry may be going away in the future, it's here for
now, so this should be there until it changes.

* data: Fixes to readme

* data: Deprecate registerX and use functions

This deprecates the following functions from the registry:

- registerReducer
- registerSelectors
- registerActions
- registerResolvers
- use

registerStore should be used instead of the above register functions,
and registerGenericStore should be used instead of plugins.

* data: docs/changelog updates for deprecations

This documents the deprecated.md and CHANGELOG.md files with
information about th wp.data registry deprecations.
It also sets the correct version for the deprecated calls to the next
minor plugin version.

* data: un-deprecate `registry.use`

This removes the deprecation of `use` because it's internally
referenced and breaks unit tests. The deprecation of `use` should be
done in a future PR that changes the behavior to not call `use`.

* data: Add deprected as script dependency to data

This adds the `wp-deprecated` script as a js script dependency to
`wp-data`.