Restore missing handbook pages after #12411 (#12452)
## Description
Following on from #12411 and #11817 it was noticed that some pages were skipped, as they weren't included in the `toc.json` file previously.

This PR includes several specific fixes which are all interconnected:
1. Adds the Internationalization page along with the Block Tutorial to the handbook hierarchy
1. An additional landing page for the Tutorials index was added
1. `Components` was added to the developers handbook alongside `Packages` and `Data Package Reference` pages.
1. In order to handle the above move, the duplicative declarations of `Packages`, `Data Package Reference` and `Components` were removed from `docs/tool/manifest.js` to prevent them adding as top-level-items. (Are these intended to be top-level-handbook-pages or deep within the developers handbook as the TOC seems to suggest?)

I also took the opportunity to convert `generate.php` from the previous PR into a JS command that allows for `docs/root-manifest.json` to be removed entirely and dynamically generated from `docs/toc.json` so it only needs updating in a single location.  That's the `getRootManifest()` and `generateRootManifestFromTOCItems()` functions which are rather messy, but work (I'm sure someone else could clean that up significantly).

## How has this been tested?
The only testing done so far is the manual review of the final `docs/manifest.json`.