Improve scrolling in fullscreen mode, notably for Edge (#13327)
In https://github.com/WordPress/gutenberg/pull/12644, we changed the behavior of scrolling for the main content area slightly, to address issues with scrolling on small screens.

As part of that, there was one small issue that made it slightly harder to scroll this main content area, only when in fullscreen mode in Microsoft Edge. This PR fixes that.

Another side effect of that prior PR was that it unset some of the scrolling rules that prevented CSS bleed (when you scroll to the end of the Block Library and proceed to scroll the body content). This PR restores that.