Update Question template for new issues (#13351)
* Update Question template for new issues

Now that Gutenberg has shipped in WordPress 5.0, we should consider updating the Question template for new issues to gently help people move their Gutenberg help requests to the main support forum as a first stop. This is because the ideal workflow is to ask for help and complete any troubleshooting needed with support first, which can then be moved to GitHub as a bug report once sufficient details are uncovered. This will help reduce duplicate requests, increase the quality of bug reports, and help developers working on the project focus on feature work and bug fixes for verified bug reports.

* Update Question template to add StackExchange

Add https://wordpress.stackexchange.com/ as a fallback suggested resource for where to get help with technical help requests.

Note: this was suggested in a Core Editor Weekly Chat on 2019-01-23 at https://wordpress.slack.com/archives/C02QB2JS7/p1548252431142200.

* Update Question template to ask to search first

I met with the WordPress Core Support Team to discuss changes to the Question template and we decided to:

* Add a line to ask people to search before posting!
* Move the handbook note up.
* Update the title to "Help Request".
* Update the description to "Please post help requests or ‘how to’ questions in support channels first".

Link to the discussion for reference: https://wordpress.slack.com/archives/C02RQC6RW/p1548349915905600

* Update Question template to move the handbook link to the end

Move the handbook note to the bottom, to be friendlier.

* Update Question template with editorial feedback

Mending the tone after an editorial review! 😍